(refactor) refactor sendMessage function in privateChatInput
