+++ client/components/privateChatInput.js
@@ -2,6 +2,7 @@ import React from 'react';
 import socket from '../sync';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router';
+import { sendPrivateMessage } from '../helpers/pchatHelpers';
 
 class privateInput extends React.Component {
   constructor(props) {
@@ -10,18 +11,7 @@ class privateInput extends React.Component {
 
   sendMessage(e) {
     if (e.which === 13 && !e.shiftKey) {
-      let msg = document.getElementById('privateMessage');
-      let msgObj = {
-        text: msg.value,
-        useroneid: global.localStorage.userid,
-        usertwoid: global.localStorage.seconduserid,
-        useronename: global.localStorage.username,
-        usertwoname: global.localStorage.secondusername,
-        chatRoom: global.localStorage.pchat
-      };
-      socket.emit('send private message', msgObj);
-      console.log('SEND MESSAGE');
-      msg.value = '';
+      sendPrivateMessage();
     }
   }
